WIP: update navbar style and refactor code
