--- src/components/Navigation/MobileNav.tsx
@@ -1,224 +1,0 @@
-import {
-  Box,
-  Button,
-  ButtonGroup,
-  Divider,
-  Drawer,
-  DrawerBody,
-  DrawerCloseButton,
-  DrawerContent,
-  DrawerHeader,
-  DrawerOverlay,
-  Flex,
-  HStack,
-  IconButton,
-  Image,
-  Stack,
-  useDisclosure
-} from '@chakra-ui/react'
-import {Link, navigate} from 'gatsby'
-import {FC} from 'react'
-import {
-  AiOutlineArrowLeft,
-  AiOutlineSearch,
-  AiOutlineShop,
-  AiOutlineShoppingCart
-} from 'react-icons/ai'
-import {FaPhoneAlt} from 'react-icons/fa'
-
-import {IoCloseOutline, IoMenuOutline} from 'react-icons/io5'
-import {useContactModal} from '../../services/contact'
-import {LayoutMode} from '../../types/commonTypes'
-import {NavAuthButton} from './NavAuthButton'
-import {BottomNavLinks} from './NavLinks'
-import {Logo} from '../../common/assets/Logo'
-
-interface IMobileNavProps {
-  mode?: LayoutMode
-  onSearchClick?: () => void
-  onBasketClick: () => void
-}
-
-export const MobileNav: FC<IMobileNavProps> = ({
-  mode,
-  onSearchClick,
-  onBasketClick
-}) => {
-  const {isOpen, onToggle} = useDisclosure()
-
-  const contactModal = useContactModal()
-
-  return (
-    <>
-      <Box
-        px="4"
-        boxShadow={mode === 'website' ? 'dark' : 'light'}
-        pos="relative"
-        bg="white"
-        zIndex="sticky">
-        <Flex h="3.75rem" justify="space-between" align="center">
-          {mode === 'website' ? (
-            <IconButton
-              aria-label="Open menu"
-              variant="ghost"
-              icon={isOpen ? <IoCloseOutline /> : <IoMenuOutline />}
-              fontSize="xl"
-              onClick={onToggle}
-            />
-          ) : (
-            <Button
-              as={Link}
-              to="/"
-              leftIcon={<AiOutlineArrowLeft />}
-              aria-label="Zur Website"
-              variant="link"
-              size="xs">
-              Zurück
-            </Button>
-          )}
-
-          <Box>
-            {/* <Image
-              onClick={() => {
-                void navigate('/')
-              }}
-              h=".875rem"
-              w="10rem"
-              src="/images/red_logo.svg"
-            /> */}
-            <Logo
-              maxWidth={{
-                base: '10rem',
-                sm: '12rem',
-                md: '15rem'
-              }}
-              height="auto"
-              // height=".875rem"
-              //width="10rem"
-              //transform="scale(0.5)"
-              objectFit="contain"
-              // display={mode === 'website' ? 'block' : 'none'}
-              cursor="pointer"
-              onClick={() => {
-                void navigate('/')
-              }}
-              color="#E3000F"
-              alt="logo"
-            />
-          </Box>
-          <HStack justifySelf="end">
-            <IconButton
-              aria-label="Kontakt"
-              size="md"
-              variant="ghost"
-              icon={<FaPhoneAlt />}
-              onClick={() => {
-                contactModal.onOpen()
-              }}
-            />
-            <NavAuthButton />
-            {mode === 'website' ? (
-              <IconButton
-                as={Link}
-                to="/products"
-                icon={<AiOutlineShop />}
-                aria-label="Shop"
-                onClick={onSearchClick}
-              />
-            ) : (
-              <ButtonGroup spacing="0">
-                <IconButton
-                  variant="ghost"
-                  icon={<AiOutlineSearch />}
-                  aria-label="Nach Artikel suchen"
-                  onClick={onSearchClick}
-                />
-                <IconButton
-                  variant="ghost"
-                  icon={<AiOutlineShoppingCart />}
-                  aria-label="Warenkorb"
-                  onClick={onBasketClick}
-                />
-              </ButtonGroup>
-            )}
-          </HStack>
-        </Flex>
-        <Drawer placement="left" onClose={onToggle} isOpen={isOpen}>
-          <DrawerOverlay />
-          <DrawerContent>
-            <DrawerCloseButton />
-            <DrawerHeader>
-              {/* <Image
-                onClick={() => {
-                  void navigate('/')
-                }}
-                h=".875rem"
-                w="10rem"
-                src="/images/red_logo.svg"
-              /> */}
-              <Logo
-                //maxWidth='20rem'
-                //height="auto"
-                height=".875rem"
-                width="10rem"
-                //transform="scale(0.5)"
-                objectFit="contain"
-                // display={mode === 'website' ? 'block' : 'none'}
-                cursor="pointer"
-                onClick={() => {
-                  void navigate('/')
-                }}
-                color="#E3000F"
-                alt="logo"
-              />
-            </DrawerHeader>
-            <Divider />
-            <DrawerBody>
-              <Stack spacing="8" pt="4">
-                <BottomNavLinks
-                  childrenTextAlign="left"
-                  w="full"
-                  px="2"
-                  zIndex="5"
-                  spacing="6"
-                  fontSize="md"
-                  onClick={onToggle}
-                />
-
-                <Stack
-                  direction="row"
-                  justify="space-between"
-                  align="center"
-                  spacing="4"
-                  mt="4">
-                  <Button
-                    w="full"
-                    variant="outline"
-                    leftIcon={<FaPhoneAlt />}
-                    onClick={() => {
-                      contactModal.onOpen({
-                        meta: {}
-                      })
-                    }}>
-                    Anfragen
-                  </Button>
-                </Stack>
-              </Stack>
-            </DrawerBody>
-          </DrawerContent>
-        </Drawer>
-      </Box>
-      <Box
-        onClick={onToggle}
-        display={isOpen ? 'block' : 'none'}
-        bg="rgba(0,0,0,0.7)"
-        h="100vh"
-        pos="absolute"
-        top="0"
-        left="0"
-        zIndex="3"
-        w="full"></Box>
-    </>
-  )
-}
-export default MobileNav
